requirements.txt
discord.py
requests
beautifulsoup4
